requirements.txt
fastapi
uvicorn[standard]
python-multipart
dataiku-api-client
pydantic
aiofiles
pandas
